requirements.txt
hydra-core
pytorch
huggingface
pandas
numpy
wandb
transformers
datasets